+++ pyproject.toml
@@ -16,7 +16,8 @@ dependencies = [
     "pydantic==1.9.0",
     "coinbasepro==0.3.0",
     "PyYaml==6.0",
-    "discord==1.7.3"
+    "discord==1.7.3",
+    "google-cloud-secret-manager==2.8.0"
 ]
 
 [project.scripts]
